requirements.txt
cassandra-driver==3.7.1
pylint==1.7.2
